+++ src/routes/entity/browse/+page.svelte
@@ -626,6 +626,7 @@
 					/>{/if}
 				<EntityFileManager
 					isOpen={isFileManagerOpen}
+					currentFilename={selectedFilename}
 					on:close={() => (isFileManagerOpen = false)}
 					on:change={async () => {
 						await loadFiles();