+++ src/routes/table/browse/+page.svelte
@@ -613,6 +613,7 @@
 					/>{/if}
 				<TableDefFileManager
 					isOpen={isFileManagerOpen}
+					currentFilename={selectedFilename}
 					on:close={() => (isFileManagerOpen = false)}
 					on:change={async () => {
 						await loadFiles();